sensor-collector/requirements.txt
aiomqtt==2.0.0
httpx==0.27.0
pydantic==2.9.2
pyyaml==6.0.2
pytest==8.3.3
pytest-asyncio==0.24.0
